split each src into seperate modules
