+++ src/middlewares/general/general.go
@@ -1,0 +1,9 @@
+package general
+
+import "github.com/gin-gonic/gin"
+
+func Ping(c *gin.Context) {
+	c.JSON(200, gin.H{
+		"message": "tradingview-go",
+	})
+}
