+++ src/middlewares/logger/logger.go
@@ -1,0 +1,25 @@
+package logger
+
+import (
+	"bytes"
+	"fmt"
+
+	"github.com/gin-gonic/gin"
+)
+
+type responseBodyWriter struct {
+	gin.ResponseWriter
+	body *bytes.Buffer
+}
+
+func (r responseBodyWriter) Write(b []byte) (int, error) {
+	r.body.Write(b)
+	return r.ResponseWriter.Write(b)
+}
+
+func LogResponseBody(c *gin.Context) {
+	w := &responseBodyWriter{body: &bytes.Buffer{}, ResponseWriter: c.Writer}
+	c.Writer = w
+	c.Next()
+	fmt.Println("Response body: " + w.body.String())
+}
